+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Search/BraveSearchManager.swift
@@ -32,8 +32,6 @@ class BraveSearchManager: NSObject {
   /// such as whether to use search fallback, should safe search be performed etc.
   private let domainCookies: [HTTPCookie]
 
-  private let profile: Profile
-
   /// The result we got from querying the fallback search engine.
   var fallbackQueryResult: String?
   /// Whether the call to the fallback search engine is pending.
@@ -65,7 +63,7 @@ class BraveSearchManager: NSObject {
 
   private var callbackLog: BraveSearchLogEntry.FallbackLogEntry?
 
-  init?(profile: Profile, url: URL, cookies: [HTTPCookie]) {
+  init?(url: URL, cookies: [HTTPCookie]) {
     if !Self.isValidURL(url) {
       return nil
     }
@@ -75,7 +73,6 @@ class BraveSearchManager: NSObject {
       let queryItem = components.valueForQuery("q")
     else { return nil }
 
-    self.profile = profile
     self.url = url
     self.query = queryItem
     self.domainCookies = cookies.filter { $0.domain == url.host }
